fix: refresh is a custom verb, not a resource
The MAAS API guidelines clearly define that path components should be
resources. And if you need a "Custom verb" in place of standard
GET/PUT/PATCH semantics, then it should use a `:` vs a `/` as the
delineator. That does make it clearer that it is an action on the
resource rather than its own separate resource.

We may still want it to just be a PATCH that just updates the
information, but at least it shouldn't be a '/'